go/consensus/cometbft: Refactor code and improve consistency

+++ go/consensus/cometbft/full/archive.go
@@ -178,7 +178,7 @@ func NewArchive(ctx context.Context, cfg ArchiveConfig) (consensusAPI.Service, e
 		},
 		Identity:            cfg.Identity,
 		DisableCheckpointer: true,
-		InitialHeight:       uint64(srv.genesisHeight),
+		InitialHeight:       srv.genesisHeight,
 		// ReadOnly should actually be preferable for archive but there is a badger issue with read-only:
 		// https://discuss.dgraph.io/t/read-only-log-truncate-required-to-run-db/16444/2
 		ReadOnlyStorage: false,
